backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.23",
    "fastapi>=0.115.6",
    "fugashi>=1.4.0",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.9",
    "langchain-openai>=0.2.11",
    "openai>=1.57.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.3.1",
    "sentencepiece>=0.2.0",
    "unidic-lite>=1.0.8",
    "uvicorn>=0.32.1",
]
